Minor Refactoring
Rename package Scheduler to Schedule.
Change comment header
Let Model implement getting the internal backing list of all Meetings
	-this is so that the UI can "listen" to display the meetings in the timetable
